Fix for new EIC Particle model

+++ trackeff/main.go
@@ -149,14 +149,14 @@ func makeTrackEffPlotters(filename string, pTMin, pTMax, fracCut, etaLimit float
 				continue
 			}
 
-			pMag := math.Sqrt(math.Pow(part.GetP().GetX(), 2) + math.Pow(part.GetP().GetY(), 2) + math.Pow(part.GetP().GetZ(), 2))
-			eta := math.Atanh(part.GetP().GetZ() / pMag)
-			pT := math.Sqrt(math.Pow(part.GetP().GetX(), 2) + math.Pow(part.GetP().GetY(), 2))
+			pMag := math.Sqrt(math.Pow(float64(part.GetP().GetX()), 2) + math.Pow(float64(part.GetP().GetY()), 2) + math.Pow(float64(part.GetP().GetZ()), 2))
+			eta := math.Atanh(float64(part.GetP().GetZ()) / pMag)
+			pT := math.Sqrt(math.Pow(float64(part.GetP().GetX()), 2) + math.Pow(float64(part.GetP().GetY()), 2))
 			chargeMag := math.Abs(float64(part.GetCharge()))
 			poqMag := pMag / chargeMag
-			diffMag := math.Sqrt(math.Pow(track.Segment[0].GetPoq().GetX()-part.GetP().GetX()/chargeMag, 2) +
-				math.Pow(track.Segment[0].GetPoq().GetY()-part.GetP().GetY()/chargeMag, 2) +
-				math.Pow(track.Segment[0].GetPoq().GetZ()-part.GetP().GetZ()/chargeMag, 2))
+			diffMag := math.Sqrt(math.Pow(track.Segment[0].GetPoq().GetX()-float64(part.GetP().GetX())/chargeMag, 2) +
+				math.Pow(track.Segment[0].GetPoq().GetY()-float64(part.GetP().GetY())/chargeMag, 2) +
+				math.Pow(track.Segment[0].GetPoq().GetZ()-float64(part.GetP().GetZ())/chargeMag, 2))
 			fracDiff := diffMag / poqMag
 
 			// cuts
@@ -177,9 +177,9 @@ func makeTrackEffPlotters(filename string, pTMin, pTMax, fracCut, etaLimit float
 				continue
 			}
 
-			pMag := math.Sqrt(math.Pow(part.GetP().GetX(), 2) + math.Pow(part.GetP().GetY(), 2) + math.Pow(part.GetP().GetZ(), 2))
-			eta := math.Atanh(part.GetP().GetZ() / pMag)
-			pT := math.Sqrt(math.Pow(part.GetP().GetX(), 2) + math.Pow(part.GetP().GetY(), 2))
+			pMag := math.Sqrt(math.Pow(float64(part.GetP().GetX()), 2) + math.Pow(float64(part.GetP().GetY()), 2) + math.Pow(float64(part.GetP().GetZ()), 2))
+			eta := math.Atanh(float64(part.GetP().GetZ()) / pMag)
+			pT := math.Sqrt(math.Pow(float64(part.GetP().GetX()), 2) + math.Pow(float64(part.GetP().GetY()), 2))
 
 			// cuts
 			if pT < pTMin || pT > pTMax {
